online_analysis_teget_error turned off
It has been noticed this line can sometimes stop the solvent branch due to the simulation reaching the error at early stage of the job, leading to a stop of solvent branching simulation even 100 iterations was defined to swap with complex simulation. 

There is no harm turning this off, the simulation will just proceed normally.